+++ docs/CHANGELOG.md
@@ -19,6 +19,7 @@
 * Item popups aren't weirdly tall on Firefox anymore.
 * Armor stats now match the order in the game.
 * Infusion calculator now always gives you the full value of your infusion.
+* Show a warning that your max light may be wrong if you have classified items.
 
 # 4.15.0